requirements.txt
anyio==3.5.0
asgiref==3.5.0
asyncio==3.4.3
bcrypt==3.2.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
cffi==1.15.0
charset-normalizer==2.0.12
click==8.0.4
dnspython==2.2.0
ecdsa==0.17.0
email-validator==1.1.3
fastapi==0.74.0
greenlet==1.1.2
h11==0.13.0
httptools==0.2.0
idna==3.3
itsdangerous==2.1.0
Jinja2==3.0.3
MarkupSafe==2.1.0
orjson==3.6.7
passlib==1.7.4
psycopg-binary==3.0.9
pyasn1==0.4.8
pycparser==2.21
pydantic==1.9.0
python-dotenv==0.19.2
python-jose==3.3.0
python-multipart==0.0.5
PyYAML==5.4.1
requests==2.27.1
rsa==4.8
six==1.16.0
sniffio==1.2.0
soupsieve==2.3.1
SQLAlchemy==1.4.31
sqlalchemy2-stubs==0.0.2a20
sqlmodel==0.0.6
starlette==0.17.1
typing_extensions==4.1.1
ujson==4.3.0
urllib3==1.26.8
uvicorn==0.15.0
uvloop==0.16.0
watchgod==0.7
websockets==10.1
wget==3.2
